Updated weightInLbs properties to double datatype throughout all layers

+++ src/main/java/walter/duncan/dndwebapi/businessmodels/EquipmentModel.java
@@ -1,7 +1,7 @@
 package walter.duncan.dndwebapi.businessmodels;
 
 public final class EquipmentModel extends GameInformationModel {
-    public EquipmentModel(String name, String description, int valueInCopperPieces, int weightInLbs) {
+    public EquipmentModel(String name, String description, int valueInCopperPieces, double weightInLbs) {
         super(name, description, valueInCopperPieces, weightInLbs);
     }
 }
